requirements.txt
smolagents>=1.13.0
gradio>=5.0.0
pytz
python-dateutil
pydantic
fastapi
uvicorn
pymongo
motor
python-dotenv
requests
duckduckgo_search
markdownify
numpy
scikit-learn
sentence-transformers
faiss-cpu
chromadb